Tweak top level Makefile and others to clean and run all examples in the new project structure
Makefile:  descend into projects to build non-framework items
platforms/getPlatform.sh: improve message
projects/assets/Project.mk: temporarily remove av component libs
projects/assets/applications/Makefile: default OCPI_LIBRARY_PATH correctly
 -- and enable running with no PATH setup
projects/assets/hdl/devices/Makefile: temporarily remove zipper devices
projects/assets/hdl/devices/ad9361_adc.test/AD9361_BIST.h: use correct errno symbol EBADF
projects/assets/hdl/devices/ad9361_adc.test/setup.mk: make it work on centos6 -- std=c++0x
projects/assets/hdl/devices/ad9361_dac.test/setup.mk: ditto
projects/assets/hdl/devices/lime_rx_proxy.test/lime_rx_proxy_test_app.cxx: fix for centos6
projects/assets/hdl/devices/lime_tx_proxy.test/lime_tx_proxy_test_app.cxx: ditto
runtime/ctests/src/run_tests.sh: av-sync
scripts/build-opencpi.sh: change to descend into projects for building
tools/cdk/include/hdl/isim.mk: only warn about xilinx paths if cleaning

+++ projects/assets/hdl/devices/lime_rx_proxy.test/lime_rx_proxy_test_app.cxx
@@ -215,10 +215,10 @@ int main(int argc, char **argv)
       //  	{
       //  	  printLimits("Error: invalid sample_rate.\n", sample_rate, sample_rate_min, sample_rate_max);
       //  	}
-      cout << "Setting sample rate to " << to_string(sample_rate) << " MHz " << endl;
+      cout << "Setting sample rate to " << to_string((long double)sample_rate) << " MHz " << endl;
 
       // Si5338:CLKIN (i.e. Si5338:CH0 = Rx clock to Lime transceiver) = 2 x sample_rate x 1 MHz
-      ClkInStr = to_string(2 * sample_rate * 1e6).c_str();
+      ClkInStr = to_string((long double)(2 * sample_rate * 1e6)).c_str();
 
       // Only required to set CH0 of the Clock Synthesis device
       clock_gen_config <<
@@ -267,10 +267,10 @@ int main(int argc, char **argv)
       //  	{
       //  	  printLimits("Error: invalid sample_rate.\n", sample_rate, sample_rate_min, sample_rate_max);
       //  	}
-      cout << "Setting sample rate to " << to_string(sample_rate) << " MHz " << endl;
+      cout << "Setting sample rate to " << to_string((long double)sample_rate) << " MHz " << endl;
 
       // Si5351:CLKIN (i.e. Si5351:CH4 & 5 = Rx clock to Lime transceiver & FPGA) = 2 x sample_rate x 1 MHz
-      rxClkInStr = to_string(2 * sample_rate * 1e6).c_str();
+      rxClkInStr = to_string((long double)(2 * sample_rate * 1e6)).c_str();
 
       // CH 4&5 tx CH 2&3 rx
       clock_gen_config <<
